feat(bench): move rayon bench and group benches to enable comparison
